Terminando meus services e controllers

+++ src/main/java/com/senai/ControleDeAcessoSpring/domain/repository/SubTurmaRepository.java
@@ -3,5 +3,5 @@ package com.senai.ControleDeAcessoSpring.domain.repository;
 import com.senai.ControleDeAcessoSpring.domain.entity.turma.SubTurma;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubTurmaRepository extends JpaRepository<SubTurma, Float> {
+public interface SubTurmaRepository extends JpaRepository<SubTurma, Long> {
 }
